WIP: Add mocha chai and test configs. Decompose util
