+++ .mocharc.yml
@@ -1,0 +1,2 @@
+require: "./test/setup.js"
+recursive: true
