+++ lib/actions/deploy/production.js
@@ -13,7 +13,9 @@ const {
   site: { basePath, redirects, excludes },
   production: { domain, bucket }
 } = config;
-const { write, getLoggers, getFiles, exec } = require("../../util");
+const { write, getLoggers } = require("../../util/log");
+const { getFiles } = require("../../util/file");
+const { exec } = require("../../util/process");
 const { Deployment } = require("../../github/deployment");
 
 const { log, error } = getLoggers("deploy:production");
