+++ lib/actions/deploy/staging.js
@@ -12,7 +12,8 @@ const SURGE_PATH = require.resolve("surge/lib/cli");
 
 const { config, validate } = require("../../config");
 const { build, site, staging: { domain } } = config;
-const { getLoggers, exec } = require("../../util");
+const { getLoggers } = require("../../util/log");
+const { exec } = require("../../util/process");
 const { Deployment } = require("../../github/deployment");
 
 const { log, error } = getLoggers("deploy:staging");
