+++ lib/actions/serve.js
@@ -11,7 +11,7 @@ const handler = require("serve-handler");
 
 const { config, validate } = require("../config");
 const { build, site } = config;
-const { getLoggers, write } = require("../util");
+const { getLoggers, write } = require("../util/log");
 
 const { error, log } = getLoggers("serve");
