+++ lib/config/index.js
@@ -2,7 +2,7 @@
 
 const path = require("path");
 const chalk = require("chalk");
-const { exists } = require("../util");
+const { exists } = require("../util/file");
 const DEFAULTS = require("./defaults");
 
 // ============================================================================
