+++ lib/github/deployment.js
@@ -6,7 +6,7 @@
 const chalk = require("chalk");
 const { Octokit } = require("@octokit/rest");
 const { config: { build, github } } = require("../config");
-const { gitInfo } = require("../util");
+const { gitInfo } = require("../util/git");
 
 const DEPLOYMENT_CREATED = 201;
